fixed download cancelling and title and author not renaming

+++ app/src/main/java/com/deniscerri/ytdlnis/work/DownloadWorker.kt
@@ -4,20 +4,18 @@ import android.app.PendingIntent
 import android.app.Service
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
+import android.os.Handler
 import android.os.Looper
-import android.os.SystemClock
-import android.provider.DocumentsContract
 import android.util.Log
 import android.widget.Toast
 import androidx.work.ForegroundInfo
 import androidx.work.Worker
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
-import com.deniscerri.ytdlnis.BuildConfig
 import com.deniscerri.ytdlnis.MainActivity
 import com.deniscerri.ytdlnis.R
 import com.deniscerri.ytdlnis.database.DBManager
+import com.deniscerri.ytdlnis.database.models.DownloadItem
 import com.deniscerri.ytdlnis.database.models.HistoryItem
 import com.deniscerri.ytdlnis.database.repository.DownloadRepository
 import com.deniscerri.ytdlnis.util.FileUtil
@@ -27,7 +25,6 @@ import com.yausername.youtubedl_android.YoutubeDLRequest
 import kotlinx.coroutines.runBlocking
 import java.io.File
 import java.util.regex.Pattern
-import kotlin.math.abs
 
 
 class DownloadWorker(
@@ -35,22 +32,35 @@ class DownloadWorker(
     workerParams: WorkerParameters
 ) : Worker(context, workerParams) {
     override fun doWork(): Result {
-        workID = inputData.getLong("workID", SystemClock.uptimeMillis())
+        itemId = inputData.getLong("id", 0)
+        if (itemId == 0L) return Result.failure()
+
         val notificationUtil = NotificationUtil(context)
         val dbManager = DBManager.getInstance(context)
         val dao = dbManager.downloadDao
         val repository = DownloadRepository(dao)
         val commandTemplateDao = dbManager.commandTemplateDao
         val historyDao = dbManager.historyDao
+        val handler = Handler(Looper.getMainLooper())
 
-        val downloadItem = repository.getDownloadByWorkID(workID)
+
+        val downloadItem: DownloadItem?
+        try {
+            downloadItem = repository.getItemByID(itemId)
+        }catch (e: Exception){
+            e.printStackTrace()
+            return Result.failure()
+        }
+
+        Log.e(TAG, downloadItem.toString())
+
         runBlocking{
-            repository.setDownloadStatus(downloadItem, DownloadRepository.status.Active)
+            repository.setDownloadStatus(downloadItem, DownloadRepository.Status.Active)
         }
         val intent = Intent(context, MainActivity::class.java)
         val pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_IMMUTABLE)
-        val notification = notificationUtil.createDownloadServiceNotification(pendingIntent, downloadItem.title, workID.toInt(), NotificationUtil.DOWNLOAD_SERVICE_CHANNEL_ID)
-        val foregroundInfo = ForegroundInfo(workID.toInt(), notification)
+        val notification = notificationUtil.createDownloadServiceNotification(pendingIntent, downloadItem.title, downloadItem.id.toInt(), NotificationUtil.DOWNLOAD_SERVICE_CHANNEL_ID)
+        val foregroundInfo = ForegroundInfo(downloadItem.id.toInt(), notification)
         setForegroundAsync(foregroundInfo)
 
         
@@ -61,7 +71,7 @@ class DownloadWorker(
 
         val tempFolder = StringBuilder(context.cacheDir.absolutePath + """/${downloadItem.title}##${downloadItem.type}""")
         tempFolder.append("##${downloadItem.format.format_id}")
-        var tempFileDir = File(tempFolder.toString())
+        val tempFileDir = File(tempFolder.toString())
         tempFileDir.delete()
         tempFileDir.mkdir()
 
@@ -89,6 +99,9 @@ class DownloadWorker(
             val filters = java.lang.String.join(",", sponsorBlockFilters)
             request.addOption("--sponsorblock-remove", filters)
         }
+
+        request.addCommands(listOf("--replace-in-metadata","title",".*.",downloadItem.title));
+        request.addCommands(listOf("--replace-in-metadata","uploader",".*.",downloadItem.author));
 
         when(type){
             "audio" -> {
@@ -171,24 +184,32 @@ class DownloadWorker(
         runCatching {
             YoutubeDL.getInstance().execute(request, downloadItem.id.toString()){ progress, _, line ->
                 setProgressAsync(workDataOf("progress" to progress.toInt()))
+                setProgressAsync(workDataOf("output" to line))
                 val title: String = downloadItem.title
                 notificationUtil.updateDownloadNotification(
-                    downloadItem.workID.toInt(),
+                    downloadItem.id.toInt(),
                     line, progress.toInt(), 0, title,
                     NotificationUtil.DOWNLOAD_SERVICE_CHANNEL_ID
                 )
             }
         }.onSuccess {
             //move file from internal to set download directory
-
-            val finalPath = moveFile(tempFileDir.absoluteFile, downloadLocation){ progress ->
-                setProgressAsync(workDataOf("progress" to progress))
+            var finalPath : String?
+            try {
+                finalPath = moveFile(tempFileDir.absoluteFile, downloadLocation){ progress ->
+                    setProgressAsync(workDataOf("progress" to progress))
+                }
+            }catch (e: Exception){
+                finalPath = context.getString(R.string.unfound_file)
+                handler.postDelayed({
+                    Toast.makeText(context, e.message, Toast.LENGTH_SHORT).show()
+                }, 1000)
             }
             //put download in history
             val incognito = sharedPreferences.getBoolean("incognito", false)
             if (!incognito) {
                 val unixtime = System.currentTimeMillis() / 1000
-                val historyItem = HistoryItem(0, downloadItem.url, downloadItem.title, downloadItem.author, downloadItem.duration, downloadItem.thumb, downloadItem.type, unixtime, finalPath, downloadItem.website, downloadItem.format)
+                val historyItem = HistoryItem(0, downloadItem.url, downloadItem.title, downloadItem.author, downloadItem.duration, downloadItem.thumb, downloadItem.type, unixtime, finalPath!!, downloadItem.website, downloadItem.format)
                 runBlocking {
                     historyDao.insert(historyItem)
                 }
@@ -197,43 +218,51 @@ class DownloadWorker(
                 dao.delete(downloadItem.id)
             }
         }.onFailure {
-            tempFileDir.delete()
-            Looper.prepare()
-            Toast.makeText(context, it.message, Toast.LENGTH_LONG).show()
-            Log.e(TAG, context.getString(R.string.failed_download), it)
-            notificationUtil.updateDownloadNotification(
-                downloadItem.workID.toInt(),
-                context.getString(R.string.failed_download), 0, 0, downloadItem.title,
-                NotificationUtil.DOWNLOAD_SERVICE_CHANNEL_ID
-            )
+            if (it is YoutubeDL.CanceledException) {
+                downloadItem.status = DownloadRepository.Status.Cancelled.toString()
+                runBlocking {
+                    dao.update(downloadItem)
+                }
+                return Result.failure()
+            }else{
+                tempFileDir.delete()
+                handler.postDelayed({
+                    Toast.makeText(context, it.message, Toast.LENGTH_SHORT).show()
+                }, 1000)
 
-            downloadItem.status = DownloadRepository.status.Errored.toString()
-            runBlocking {
-                dao.update(downloadItem)
+                Log.e(TAG, context.getString(R.string.failed_download), it)
+                notificationUtil.updateDownloadNotification(
+                    downloadItem.id.toInt(),
+                    context.getString(R.string.failed_download), 0, 0, downloadItem.title,
+                    NotificationUtil.DOWNLOAD_SERVICE_CHANNEL_ID
+                )
+
+                downloadItem.status = DownloadRepository.Status.Error.toString()
+                runBlocking {
+                    dao.update(downloadItem)
+                }
+                return Result.failure()
             }
-            return Result.failure()
         }
         
         return Result.success()
     }
 
-    private fun getDownloadLocation(type: String, context: Context): String? {
-        val sharedPreferences = context.getSharedPreferences("root_preferences",
-            Service.MODE_PRIVATE
-        )
-        val downloadsDir: String? = if (type == "audio") {
-            sharedPreferences.getString("music_path", context.getString(R.string.music_path))
-        } else {
-            sharedPreferences.getString("video_path", context.getString(R.string.video_path))
-        }
-        return downloadsDir
-    }
+//    private fun getDownloadLocation(type: String, context: Context): String? {
+//        val sharedPreferences = context.getSharedPreferences("root_preferences",
+//            Service.MODE_PRIVATE
+//        )
+//        val downloadsDir: String? = if (type == "audio") {
+//            sharedPreferences.getString("music_path", context.getString(R.string.music_path))
+//        } else {
+//            sharedPreferences.getString("video_path", context.getString(R.string.video_path))
+//        }
+//        return downloadsDir
+//    }
 
 
+    @Throws(Exception::class)
     private fun moveFile(originDir: File, downLocation: String, progress: (progress: Int) -> Unit) : String{
-        val destDir = Uri.parse(downLocation).run {
-            DocumentsContract.buildChildDocumentsUriUsingTree(this, DocumentsContract.getTreeDocumentId(this))
-        }
         val fileUtil = FileUtil()
         val path = fileUtil.moveFile(originDir, context, downLocation){ p ->
             progress(p)
@@ -242,7 +271,7 @@ class DownloadWorker(
     }
 
     companion object {
-        var workID: Long = 0
+        var itemId: Long = 0
         const val TAG = "DownloadWorker"
     }
